+++ compiler/GHC/Data/Bag.hs
@@ -7,6 +7,7 @@ Bag: an unordered collection with duplicates
 -}
 
 {-# LANGUAGE ScopedTypeVariables, DeriveTraversable, TypeFamilies #-}
+{-# OPTIONS_GHC -Wno-deprecations #-}
 
 module GHC.Data.Bag (
         Bag, -- abstract type