+++ hadrian/src/Rules/Dependencies.hs
@@ -1,3 +1,5 @@
+{-# OPTIONS_GHC -Wno-deprecations #-}
+
 module Rules.Dependencies (buildPackageDependencies) where
 
 import Data.Bifunctor